Refactoring the Default Styles
