--- app/components/AppText/AppText.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native'
-
-import styles from './styles';
-
-export default function AppText({ children, style }) {
-    return (
-        <Text style={[styles.text, style]}>{children}</Text>
-    )
-}
-
-
