--- app/components/AppText/styles.js
@@ -1,10 +1,0 @@
-import { StyleSheet, Platform } from 'react-native'
-
-const styles = StyleSheet.create({
-    text: {
-        fontSize: 18,
-        fontFamily: Platform.OS === "android" ? "Roboto" : "Avenir"
-    }
-})
-
-export default styles;
