+++ app/components/AppTextInput.js
@@ -3,6 +3,7 @@ import { View, StyleSheet, TextInput, Platform } from 'react-native';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 
 import colors from '../config/colors'
+import defaultStyles from '../config/styles'
 
 export default function AppTextInput({ icon, ...otherProps }) {
   return (
@@ -10,10 +11,10 @@ export default function AppTextInput({ icon, ...otherProps }) {
       {icon && <MaterialCommunityIcons 
       name={icon} 
       size={20} 
-      color={colors.medium} 
+      color={defaultStyles.colors.medium} 
       style={styles.icon}
       />}
-      <TextInput style={styles.textInput} {...otherProps} />
+      <TextInput style={defaultStyles.text} {...otherProps} />
      </View>
   );
 }
@@ -31,8 +32,6 @@ const styles = StyleSheet.create({
     marginRight: 10,
   },
   textInput: {
-    fontSize: 18,
-    fontFamily: Platform.OS === "android" ? "Roboto" : "Avenir",
-    color: colors.dark,
+    
   },
 })
