+++ app/components/Card.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { View, StyleSheet, Image } from 'react-native'
 import Colors from '../config/colors'
-import AppText from './AppText/AppText'
+import AppText from './AppText'
 
 export default function Card({ title, subTitle, pic }) {
     return (
