+++ app/components/ListItem.js
@@ -2,7 +2,7 @@ import React from "react";
 import { View, StyleSheet, Image, TouchableHighlight } from "react-native";
 import Swipeable from "react-native-gesture-handler/Swipeable";
 
-import AppText from "./AppText/AppText";
+import AppText from "./AppText";
 import colors from "../config/colors";
 
 export default function ListItem({
