+++ app/config/colors.js
@@ -1,4 +1,5 @@
-export default { 
+export default 
+{
   primary: '#fc5c65',
   secondary: '#4ecdc4', 
   black: '#000',
