+++ app/config/styles.js
@@ -1,0 +1,12 @@
+import { StyleSheet, Platform } from 'react-native'
+
+import colors from './colors'
+
+export default {
+    colors,
+    text: {
+        color: colors.dark,
+        fontSize: 18,
+        fontFamily: Platform.OS === "android" ? "Roboto" : "Avenir",
+    },
+}
