+++ app/screens/ListingDetailsScreen.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { View, Image, StyleSheet } from 'react-native'
 
 import ListItem from '../components/ListItem'
-import AppText from '../components/AppText/AppText'
+import AppText from '../components/AppText'
 import  Colors  from '../config/colors'
 
 export default function ListingDetailsScreen() {
